==> R/01_data_load.qmd <==
---
format: html
editor: visual
---

## Load data

### Downloading data

```{r}
#| output: false
#| warning: false
#| echo: false
library("here")
library("tidyverse")
source("99_proj_func.R")
set.seed(28)
```

```{r}
#| output: false
raw_dir <- here("data/_raw/")
data_loc <- "https://gdc-hub.s3.us-east-1.amazonaws.com/download/"
download_file(raw_dir, 
              data_loc, 
              data_file = "TCGA-BRCA.star_fpkm.tsv.gz", 
              new_file_name = "gene_expression_fpkm.tsv.gz")
```

```{r}
#| output: false
download_file(raw_dir, 
              data_loc, 
              data_file = "TCGA-BRCA.clinical.tsv.gz", 
              new_file_name = "phenotypes.tsv.gz")
```

## Loading data

```{r}
#| output: false
#| warning: false
#| echo: false
gene_expression <- read_tsv(file = here("data/_raw/gene_expression_fpkm.tsv.gz"), 
                            na = c("NA",""," "))
phenotypes <- read_tsv(file = here("data/_raw/phenotypes.tsv.gz"), 
                       na = c("NA","", " "))
```

### Phenotype data

```{r}
phenotypes_subset <- phenotypes |> 
  select(sample, 
         gender.demographic,
         ajcc_pathologic_stage.diagnoses,
         age_at_diagnosis.diagnoses, 
         ethnicity.demographic, 
         days_to_death.demographic, 
         vital_status.demographic,
         treatment_or_therapy.treatments.diagnoses,
         primary_diagnosis.diagnoses, 
         prior_malignancy.diagnoses, 
         year_of_diagnosis.diagnoses)
```

### Gene expression data

Sampling 2000 random genes

```{r}
gene_expression_subset <- gene_expression |> 
  sample_n(size = 2000)
```

Pivoting longer

```{r}
gene_expression_subset_long_format <- gene_expression_subset |> 
  pivot_longer(cols = !Ensembl_ID, 
               names_to = "sample", 
               values_to = "gene_expression")

```

Joining the two datasets

```{r}
loaded_data <- gene_expression_subset_long_format |> 
  inner_join(phenotypes_subset, 
             join_by(sample))

loaded_data
```

Save the data as a tsv again

```{r}
#| output: false
loaded_data |> 
  write_tsv(file = here("data/01_dat_load.tsv.gz"), 
            na = "NA")
```
