==> R/02_clean.qmd <==
---
format: html
editor: visual
---

```{r}
#| output: false
#| warning: false
#| echo: false
library("here")
library("tidyverse")
```

```{r}
#| output: false
#| warning: false
#| echo: false
loaded_data <- read_tsv(file = here("data/01_dat_load.tsv.gz"), 
                na = "NA",  guess_max = 1001)
```

## Cleaning the data

### Renaming the columns

```{r}
clean_data <- loaded_data |> 
  rename_with(~ str_replace(., "\\..*", "")) |> 
  rename("ensembl_id" = "Ensembl_ID", 
         "cancer_stage" = "ajcc_pathologic_stage")

```

### Tidying treatment and therapy

```{r}

clean_data <- clean_data |> 
  mutate(treatment_or_therapy = str_remove_all(treatment_or_therapy, 
                                               pattern = "[\\[\\]\\']")) |> 
  separate_wider_delim(treatment_or_therapy,
                       delim = ", ",
                       names = c("treatment", "therapy"))
```

### Saving the clean data

```{r}
#| output: false
write_tsv(x = clean_data, 
          file = here("data/02_dat_clean.tsv.gz"), na = "NA")
```
