==> R/03_augment.qmd <==
---
format: html
editor: visual
---

```{r}
#| output: false
#| warning: false
#| echo: false
library("tidyverse")
library("here")
```

```{r}
#| output: false
#| warning: false
#| echo: false
clean_data <- read_tsv(file = here("data/02_dat_clean.tsv.gz"), 
                       na = "NA", 
                       guess_max = 1001)
```

## Augmenting the data

```{r}
#| output: false
#| warning: false
#| echo: false
clean_data |> 
  select(cancer_stage) |> 
  unique()
```

### Age at diagnosis from days to years

```{r}
clean_data <- clean_data  |> 
  mutate(age_at_diagnosis_years = age_at_diagnosis/365.25)
```

### Age grouping

```{r}
clean_data = clean_data |> 
  mutate(age_group = case_when(20 < age_at_diagnosis_years & age_at_diagnosis_years <= 30 ~ "(20,30]",
                               30 < age_at_diagnosis_years & age_at_diagnosis_years <= 40 ~ "(30,40]",
                               40 < age_at_diagnosis_years & age_at_diagnosis_years <= 50 ~ "(40,50]",
                               50 < age_at_diagnosis_years & age_at_diagnosis_years <= 60 ~ "(50,60]",
                               60 < age_at_diagnosis_years & age_at_diagnosis_years <= 70 ~ "(60,70]",
                               70 < age_at_diagnosis_years & age_at_diagnosis_years <= 80 ~ "(70,80]",
                               80 < age_at_diagnosis_years & age_at_diagnosis_years <= 90 ~ "(80,90]"))
```

### Years from diagnosis to death

```{r}
clean_data <- clean_data  |> 
  mutate(years_from_diagnosis_to_death = (days_to_death/365.25))
```

### Age at death in years

```{r}
clean_data <- clean_data  |> 
  mutate(age_at_death_years = age_at_diagnosis_years + years_from_diagnosis_to_death)
```

### Year of death

```{r}
clean_data <- clean_data  |> 
  mutate(year_of_death = year_of_diagnosis+years_from_diagnosis_to_death)
```

### Write to tsv

```{r}
#| output: false
clean_data |> write_tsv(file = here("data/03_dat_aug.tsv.gz"), na = "NA")
```
