==> R/04_describe.qmd <==
---
format: html
editor: visual
---

## Describing the data

```{r}
#| output: false
#| warning: false
#| echo: false
library("tidyverse")
library("table1")
library("dplyr")
library("here")
library("ggridges")
library("patchwork")
library("viridis")
```

```{r}
#| output: false
#| echo: false
#| warning: false
data <- read_tsv(file = here("data/03_dat_aug.tsv.gz"), 
                 na = "NA", 
                 guess_max = 1001
                 )
```

### Creating a table1 to summarize data

```{r}
#| warning: false
data |>
  filter(!is.na(vital_status)) |>
  table1(x = formula(~ gender + ethnicity + age_group + cancer_stage | vital_status),
         data = _)

```

### Descriptive graphs

### The below graph shows the amount of people for stage of cancer

```{r}
#| warning: false
people_per_stage <- ggplot(data, 
       aes(y = cancer_stage, 
           fill = cancer_stage)) +
  geom_bar(stat = "count", alpha=0.7) +
  geom_text(stat = "count", aes(label = after_stat(count)), 
            hjust = -0.1, size = 3.5, color = "black") +
  scale_x_continuous(limits = c(0, 900000), 
                     expand = c(0, 0)) +
  scale_fill_viridis_d(option = "plasma", 
                       na.value = "yellow3") +
  labs(
    title = "Number of People per Cancer Stage",
    caption = "Data from BRCA at: https://xenabrowser.net/",
    x = "Count",
    y = "Cancer Stage"
  ) +
  theme_minimal(base_size = 12) + 
  theme(
    plot.title = element_text(hjust = 0.5, size = 16, face = "bold"), 
    axis.text.x = element_text(hjust = 1, vjust = 1, face = "bold"), 
    axis.text.y = element_text(face = "bold"),  
    panel.grid = element_blank(),
    axis.line = element_line(color = "black", size = 0.1), 
    legend.position = "none"
    )


people_per_stage

```

```{r}
#| output: false
#| echo: false
#| warning: false
ggsave("../results/data_description/people_per_stage.png")
```

Here, it becomes clear that the most prominent cancer stage represented in the data is stage IIA, while Stage II and Stage IB are more sparse.

### Illustrating the sexes and ethnicities of the people tested (removed NA values)

```{r}
#| warning: false
sex_and_ethnicity_data <- data |> 
  filter(!is.na(gender) & !is.na(ethnicity)) 
```

```{r}
#| warning: false
gender_ethnicity_overview <- ggplot(sex_and_ethnicity_data,
       aes(x = gender, 
           fill = ethnicity))+
  geom_bar(position = "fill", alpha = 0.7) +
  scale_fill_manual(values = c("hispanic or latino" = "lightgreen", 
                               "not hispanic or latino" = "deepskyblue3", 
                               "not reported" = "hotpink")) +
  labs(
    title = "Gender and Ethnicity Overview",
    x = "Gender",
    y = "Proportion", 
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = "Ethnicity"
  ) +
  theme_minimal()+
  theme(
  plot.title = element_text(hjust = 0.5, size = 16, face = "bold"), 
  axis.title = element_text(face="bold"),
  axis.text.x = element_text(size = 12), 
  axis.text.y = element_text(face = "bold"),  
  panel.grid = element_blank(),
  axis.line = element_line(color = "black", linewidth = 0.1), 
  legend.position = "right"
  )

gender_ethnicity_overview

```

It is hereby noticiable that the majority of the ethnicities represented are categorised as "not hispanic or latino". This is a lack within in the dataset, which does not specify specific ethnicities, but rather categorizes patient as either latino or not latino.

It should be noted that the y-axis here is proportion, and that the ethnicities are represented as a fraction of the data within each gender. This is important so as to avoid any confusion on how well each gender is represented in the data (as seen in the table1 above, men only make up 1.1% of the dataset).

### Looking at whether therapy and treatment have any effect on days to death

```{r}
#| warning: false

treatment_graph_data <- data |>
    filter(!is.na(treatment), !is.na(days_to_death), treatment != "not reported")
  
therapy_graph_data  <- data |>
    filter(!is.na(therapy), !is.na(days_to_death), therapy != "not reported")

```

```{r}
#| warning: false
treatment_effects <- ggplot(treatment_graph_data, 
             aes(x = treatment, 
                 y = days_to_death, 
                 fill = treatment)) +
  geom_boxplot(alpha=0.7) +
  scale_fill_manual(values = c("yes" = "darkolivegreen", 
                             "no" = "coral3")) +
  labs(
    title = "Days to Death by Treatment",
    x = "Treatment",
    y = "Days to Death", 
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = NULL
  ) +
  theme_minimal(base_size = 10) +
  theme(
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    axis.text.x = element_text(size=11),
    strip.text = element_text(size = 12, face = "bold"),
    legend.position = "none" 
  )


```

```{r}
#| warning: false
therapy_effects <- ggplot(therapy_graph_data, 
             aes(x = therapy, 
                 y = days_to_death, 
                 fill = therapy)) +
  geom_boxplot(alpha=0.7) +
  scale_fill_manual(values = c("yes" = "darkolivegreen", 
                             "no" = "coral3")) +
  labs(
    title = "Days to Death by Therapy",
    x = "Therapy",
    y = "Days to Death",
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = NULL
  ) +
  theme_minimal(base_size = 10) +
  theme(
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    axis.text.x = element_text(size=11), 
    strip.text = element_text(size = 12, face = "bold"),
    legend.position = "none" 
  )


combined_plot <- treatment_effects / therapy_effects


combined_plot
```

```{r}
#| output: false
#| echo: false
#| warning: false
ggsave("../results/data_description/treatment_vs_therapy_v1.png")

```

This graph shows slight differences, but it is not very clear whether there is a real effect of each treatment. Therefore, it would be beneficial to represent this data in a different way, as seen below. Here, we can create a new column for the treatment combinations, and stratify this data by vitality status.

### Looking at whether therapy, treatment, both or neither have any effect on whether the patient lives or not

```{r}
#| warning: false
combined_data <- data |>
  mutate(
    combined_category = case_when(
      therapy == "yes" & treatment == "yes" ~ "Both",
      therapy == "no" & treatment == "yes" ~ "Treatment Only",
      therapy == "yes" & treatment == "no" ~ "Therapy Only",
      therapy == "no" & treatment == "no" ~ "Neither",
      TRUE ~ NA_character_
    )
  ) |>
  filter(!is.na(combined_category), !is.na(vital_status))

```

```{r}
#| warning: false
treatment_vs_therapy <- ggplot(combined_data, 
       aes(x = combined_category, 
           fill = vital_status)) +
  geom_bar(position = "fill", alpha=0.7) +
  scale_fill_manual(values = c("Alive" = "darkseagreen", "Dead" = "maroon")) +
    scale_y_continuous(labels = scales::percent) +
  labs(
    title = "Vital Status by Therapy and Treatment Combination",
    x = "Therapy and Treatment Combination",
    y = "Proportion",
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = "Vital Status"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(size=12),
    axis.text.y = element_text(face = "bold"),
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    strip.text = element_text(size = 12, face = "bold"),
    panel.grid = element_blank(), 
    axis.line = element_line(color = "black")
  )
  

treatment_vs_therapy

```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/treatment_vs_therapy_v2.png")
```

Here we see that, as common sense would dictate, there is a larger percentage of patients who die when they do not receive any treatment, and it appears that therapy is the most effective of the two.

### Graph showing whether a given age is more likely to survive

```{r}
#| warning: false
survival_by_age <- ggplot(data |> 
         filter(!is.na(age_at_diagnosis_years)), 
         aes(x = age_at_diagnosis_years, 
             fill = vital_status)) +
  geom_density(alpha = 0.3, 
               position = "identity", 
               adjust = 3) +  # adjust the smoothness
  scale_fill_manual(values = c("Alive" = "lightgreen", 
                               "Dead" = "maroon")) +
  labs(
    title = "Vital Status by Age",
    x = "Age at Diagnosis",
    y = "Density", 
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = "Vital Status"
  ) +
  theme_minimal()+
  theme(
  axis.text.y = element_text(face = "bold"),
  plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
  strip.text = element_text(size = 12, face = "bold"),
  legend.position = "top", 
  panel.grid = element_blank(), 
  axis.line = element_line(color = "black")
)


survival_by_age


```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/survival_by_age.png")
```

This graph illustrates that the majority of people survive when diagnosed earlier in life (i.e. younger than 70 years old), while the majority of people who are diagnosed later do not survive.

### Percentage of dead and alive people of each gender for each ethnicity.

```{r}
#| warning: false
data_summary <- data |>
  filter(!is.na(gender) & !is.na(vital_status) & !is.na(ethnicity) & ethnicity != "not reported") |>
  group_by(gender, ethnicity, vital_status) |>
  summarise(count = n(), .groups = "drop") |>
  group_by(ethnicity, gender) |>
  
  mutate(total_count = sum(count)) |>
  
  ungroup() |>
  
  mutate(percentage = (count / total_count) * 100) |>
  mutate(percentage = ifelse(gender == "female",
                               percentage,
                               -1*percentage)) |>

  bind_rows(tibble(gender = "male", ethnicity = "hispanic or latino", vital_status = "Dead", percentage = 0))

```

```{r}
#| warning: false
gender_ethnicity_vitality <- ggplot(data_summary, 
       aes(x = ethnicity, 
           y = ifelse(gender == "female", percentage, -percentage), 
           fill = vital_status)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.7, alpha=0.7) + 
  geom_hline(yintercept = 0, color = "black", linewidth = 0.5) +
  coord_flip() +
  scale_fill_manual(values = c("Alive" = "#66c2a5", "Dead" = "#fc8d62")) + 
  facet_wrap(~ gender, scales = "free_x", ncol = 2) + 
  labs(
    title = "Vitality by Ethnicity and Gender",
    x = NULL,
    y = "Percentage",
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = "Vital Status"
  ) +
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1),
    axis.text.y = element_text(face = "bold"),
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    strip.text = element_text(size = 12, face = "bold"),
    legend.position = "top"
  )


gender_ethnicity_vitality
```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/gender_ethnicity_vitality.png")

```

Again, it should be noted that this data is shown as a percentage, as a very small subset of the data represents men. However, it can still be noted that a greater percentage of non-hispanic people in the study have died, as opposed to latino patients. This is true for both genders, and may explain why the researchers have divided the patient ethnicities into these two groups.

### Investigating whether prior malignancy affects days to death

```{r}
#| warning: false
days_to_death_prior_malignancy <- ggplot(data |> 
         filter(!is.na(prior_malignancy) & !is.na(days_to_death)), 
       aes(x = prior_malignancy, y = days_to_death, fill = prior_malignancy)) +
  geom_violin(alpha=0.8) +  # Use violin plot
  scale_fill_manual(values = c("no" = "navyblue", "yes" = "maroon")) +
  labs(
    title = "Days to Death by Prior Malignancy",
    caption = "Data from BRCA at: https://xenabrowser.net/",
    x = "Prior Malignancy",
    y = "Days to Death"
  ) +
  theme_minimal() +
  theme(
  axis.text.x = element_text(angle = 45, hjust = 1),
  axis.text.y = element_text(face = "bold"),
  plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
  strip.text = element_text(size = 12, face = "bold"),
  legend.position = "none"
)

# display 
days_to_death_prior_malignancy

```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/days_to_death_prior_malignancy.png")
```

Here, it is clear that prior malignancy has an effect on the survival of each patient. Again, this data follows logic and shows that the presence of a prior malignancy decreases the number of days each patient has left to live.

### The below graphs shows the vitality for each type of cancer.

First the category titles are too long to be legible, so we rename them to be shorter (this does affect the understanding a little, but is worth it to be allow the viewer to read the text):

```{r}
#| output: false
#| echo: false
#| warning: false
data_primary_diagnosis_graph <- data |>
  mutate(primary_diagnosis_short = recode(primary_diagnosis,
    "Infiltrating duct carcinoma, NOS" = "Infiltrating duct",
    "Infiltrating duct mixed with other types of carcinoma" = "Infiltrating duct mixed",
    "Lobular carcinoma, NOS" = "Lobular",
    "Medullary carcinoma, NOS" = "Medullary",
    "Large cell neuroendocrine carcinoma" = "Large cell neuroendocrine",
    "Infiltrating lobular mixed with other types of carcinoma" = "Infiltrating lobular mixed",
    "Intraductal papillary adenocarcinoma with invasion" = "Intraductal papillary adeno",
    "Mucinous adenocarcinoma" = "Mucinous adeno",
    "Phyllodes tumor, malignant" = "Phyllodes tumor",
    "Paget disease and infiltrating duct carcinoma of breast" = "Paget disease and duct",
    "Apocrine adenocarcinoma" = "Apocrine adeno",
    "Infiltrating duct and lobular carcinoma" = "Duct and lobular",
    "Secretory carcinoma of breast" = "Secretory",
    "Cribriform carcinoma, NOS" = "Cribriform",
    "Metaplastic carcinoma, NOS" = "Metaplastic",
    "Pleomorphic carcinoma" = "Pleomorphic",
    "Intraductal micropapillary carcinoma" = "Intraductal micropapillary",
    "Adenoid cystic carcinoma" = "Adenoid cystic",
    "Carcinoma, NOS" = "Carcinoma",
    "Papillary carcinoma, NOS" = "Papillary",
    "Basal cell carcinoma, NOS" = "Basal cell",
    "Tubular adenocarcinoma" = "Tubular adeno"
  )) |>
    
  filter(!is.na(primary_diagnosis_short) & !is.na(vital_status))

```

```{r}
#| warning: false
cancer_type_vitality_v1 <-ggplot(data_primary_diagnosis_graph,
       aes(x = vital_status, fill = vital_status)) +
  geom_bar(alpha=0.7) +
  scale_fill_manual(values = c("Alive" = "lightgreen", "Dead" = "maroon")) +
  facet_wrap(~ primary_diagnosis_short, scales = "free_y") +
  scale_y_continuous(breaks = NULL) +
  labs(title = "Vital Status by Primary Diagnosis",
       x = NULL,
       y = NULL, 
       caption = "Data from BRCA at: https://xenabrowser.net/",
       fill = NULL) +
  theme_classic(base_size = 10) +
  theme(
    strip.text = element_text(size = 6, angle = 0, hjust = 0.5, vjust = 0.5), 
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    legend.position = "top"
  )

cancer_type_vitality_v1

```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/cancer_type_vitality_v1.png")
```

This plot shows the vitality data for each type of cancer, but is quite overwhelming to look at. Therefore, we create another graph which represents the same data as a stacked bar graph with proportions.

```{r}
#| warning: false
cancer_type_vitality_v2 <- ggplot(data_primary_diagnosis_graph,
       aes(x = primary_diagnosis_short, 
           fill = vital_status)) +
  geom_bar(position = "fill", alpha=0.7) +
  scale_fill_manual(values = c("Alive" = "lightgreen", "Dead" = "maroon")) +
  labs(title = "Proportion of Vital Status by Primary Diagnosis", 
       x = "Primary Diagnosis", 
       y = "Proportion", 
       caption = "Data from BRCA at: https://xenabrowser.net/",
       fill = "Vital Status")+
  theme_minimal() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1), 
    plot.title = element_text(hjust = 0.5, size = 14, face = "bold"),
    ) 

cancer_type_vitality_v2


```

```{r}
#| output: false
#| echo: false
#| warning: false

ggsave("../results/data_description/cancer_type_vitality_v2.png")

```

This graph is easier to interpret, and shows that some primary diagnoses (e.g. secretory breast carcinoma) have a 100% mortality rate, while others (e.g. adenoid cystic carcinoma) have much better changes of survival.
