==> R/05_analysis_1.qmd <==
---
format: html
editor: visual
---

## Analysing impact of cancer stage

```{r}
#| output: false
#| warning: false
#| echo: false
library("tidyverse")
library("here")
```

```{r}
#| output: false
#| warning: false
#| echo: false
data <- read_tsv(file = here("data/03_dat_aug.tsv.gz"), 
                       na = "NA", 
                      guess_max = 1001)
```

### Add factors to relevant colums

```{r}
data <- data |> 
  mutate(vital_status = factor(vital_status,
                               levels = c("Dead", 
                                          "Alive")),
         cancer_stage = factor(cancer_stage,
                               levels = c("Stage I", 
                                          "Stage IA", 
                                          "Stage IB", 
                                          "Stage II", 
                                          "Stage IIA", 
                                          "Stage IIB", 
                                          "Stage IIIA", 
                                          "Stage IIIB", 
                                          "Stage IIIC", 
                                          "Stage IV"))
         )
```

## Survival time by cancer stage

Let's take a look at survival times for the different cancer stages.

```{r}
#| warning: false
data |> 
  filter(cancer_stage != "Stage IB" & 
           cancer_stage != "Stage II" &
           vital_status == "Dead") |>
  ggplot(mapping = aes(x = cancer_stage, 
                       y = years_from_diagnosis_to_death,
                       fill = cancer_stage)
         ) +
  geom_boxplot() +
  labs(x = "Cancer stages",
       y = "Years until death",
       caption = "Data from BRCA at: https://xenabrowser.net/",
       title = "Years Until Death by Cancer Stage") +
  theme_minimal() +
  theme(legend.position = 'none')
```

In general it seems the worse the cancer stage of a patient is when diagnosed, the shorter until the patient dies. There is however a weird insight, namely that patients diagnosed with stage IIB cancer lives longer than patients with stage IIA cancer, which is counter intuitive. This could be an artifact of the fact that this study is right-censored.

```{r}
#| output: false
#| echo: false
#| warning: false
ggsave(file = here("results/05_01_years_until_death_by_cancer_stage.png"))
```

Let's take a look at the amount of dead and alive patients across cancer stage

```{r}
data |> 
  drop_na(cancer_stage) |>
  filter(cancer_stage != "Stage IB" & cancer_stage != "Stage II") |>
  ggplot(aes(x = cancer_stage, fill = vital_status)) + 
  geom_bar(position = "dodge") +
  scale_fill_manual(values = c("Alive" = "lightgreen", "Dead" = "maroon")) +
  labs(
    title = "Vital Status by Cancer Stage",
    x = "Cancer Stage",
    y = "Count", 
    caption = "Data from BRCA at: https://xenabrowser.net/",
    fill = "Vital Status"
  ) +
  theme_minimal()

```

We can see that most patients are diagnosed in stage IIA and stage IIB, which is also the stages with the most patients surviving.

```{r}
#| output: false
#| echo: false
#| warning: false
ggsave(file = here("results/05_02_vital_status_by_cancer_stage.png"))
```
